Split up content script into ExtensionMessageRouter
- To make it testable
